requirements.txt
asgiref==3.5.0
astroid==2.11.2
backports.zoneinfo==0.2.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
colorama==0.4.4
cryptography==36.0.2
defusedxml==0.7.1
dill==0.3.4
Django==4.0.3
django-allauth==0.50.0
django-cors-headers==3.11.0
django-rest-auth==0.9.5
djangorestframework==3.13.1
gunicorn==20.1.0
idna==3.3
isort==5.10.1
lazy-object-proxy==1.7.1
mccabe==0.7.0
oauthlib==3.2.0
Pillow==9.0.1
platformdirs==2.5.1
psycopg==3.0.11
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyJWT==2.3.0
pylint==2.13.3
python3-openid==3.2.0
pytz==2022.1
requests==2.27.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
typing-extensions==4.1.1
tzdata==2022.1
urllib3==1.26.9
whitenoise==6.0.0
wrapt==1.14.0
